Fix crashing bug when canceling TBB computations involving a reused CholeskyFlowgraph
Previously, CholeskyFlowgraph::g was (implicitly) binding to a temporary
taskgroup context created by the enclosing task groups. When the graph
was reused for a subsequent factorization, this temporary context had
been destroyed, leading TBB to chase invalid pointers when propagating
cancellation state through the task group forest.

The crash could be fixed by running the flowgraph in an isolated
context, intead we introduce a single context for the full
Cholesky factorization to simplify cancellation (i.e., expedite
termination when nonpositive definiteness is detected).