Fix intermittent crash in BlockRightLooking.
When `CholeskyFlowGraph` encounters a nonpositive pivot, it immediately
cancels execution of the global `tbb_ctx`.
This means that any child supernode c whose processing has not yet started
will never get processed. Normally this is OK because c's parent
usually becomes aware of the factorization failure before attempting
to access c's empty Schur complement. In fact, we expected the following
lines to notify the parent when the task in `tg` reponsible for processing c
did not run to completion:

      auto status = tg.wait();
      if (status != tbb::task_group_status::complete)
          shared_state->setFailed();

Unfortunately, this apparently does not work: `tg` can exit
early due to failures on a different thread but still report a
`complete` status. And more unfortunately, the thread
where the dense Cholesky factorization failed
does not always execute `shared_state->setFailed()` early enough
for parents being processed on other threads to be aware in time.
Ultimately this meant that the code attempted to access uninitialized
child Schur complements despite the safeguard:

  if (!shared_state->hasFailed()) {
     // Uninitialized Schur complements were accessed here :(
  }

To fix this, we now ensure `CholeskyFlowGraph` executes
`shared_state->setFailed()` *before* cancelling the `task_group_context`.
This is done via the new `failureCallback` notification mechanism.